fix include errors

+++ include/ab_string.h
@@ -1,6 +1,8 @@
 #ifndef AB_STRING_H
 #define AB_STRING_H
 
-unsigned long ab_strlen(const char *s);
+#include "ab_types.h"
+
+size_t ab_strlen(const char *s);
 
 #endif
